Added new future textures + created materials + created terrain prefab
The wall prefab was assigned a new material because it already belongs in the future world. Three wall (box) textures were added, and three materials created, so we might want to assign them randomly (like what happens with the player) or create separate prefabs. A new terrain prefab was created.