Future skin (#183)
* Added new future textures + created materials + created terrain prefab

The wall prefab was assigned a new material because it already belongs in the future world. Three wall (box) textures were added, and three materials created, so we might want to assign them randomly (like what happens with the player) or create separate prefabs. A new terrain prefab was created.

* Added new wall model and created new prefab

On the previous commit I'd only changed the material, but it has to be a new prefab because the model's UVs are unwrapped differently and match the new textures

* Fixed bugs, placed and changed grid, dinamically resized the terrain's texture to match the grid with patterns

* Fixed test